requirements.txt
httpx
python-dotenv
PyJWT
requests
cryptography
pytest